First commit for support of podcast streaming to upnp/dlna media renderers
The feature is still in an early stage however working with a couple of devices.

In the menu of the MediaplayerActivity there is an additional option 'select media renderer'. On selection, a dialog is shown which displays all suitable upnp/dlna devices. If a new device is picked, current playback is stopped and diverted to the selected device. If possible, playback starts from the last position.

A device's volume can be altered on the MediaplayerActivity by using the hardware volume buttons.

If permitted by the device an arbitrary playback position can be chosen as for the local player.

This feature uses the cling library http://4thline.org/projects/cling/